PrepararAmbiente/requirements.txt
conda==4.3.16
mysql-connector-python==8.0.26
openpyxl==3.0.8
pandas==1.3.3
pip-chill==1.0.1
pipenv==2021.5.29
